refactor(backends): extract OpenMM term records and unit converters from openmm.py
Phase 2 of the repo refactor (research finding 4.4). Pure, mechanical
code-move with zero behavior change: relocate two self-contained private
blocks out of q2mm/backends/mm/openmm.py into two new private sibling
modules.

- q2mm/backends/mm/_openmm_terms.py: the 7 private term dataclasses
  (_BondTerm, _AngleTerm, _VdwTerm, _Exception14, _TorsionTerm, _UBTerm,
  _CmapTerm), moved verbatim.
- q2mm/backends/mm/_openmm_units.py: the 7 scalar unit-converter wrappers
  (_bond_k_to_openmm, _angle_k_to_openmm, _bond_k_to_harmonic,
  _angle_k_to_harmonic, _vdw_sigma_nm, _vdw_radius_to_openmm,
  _vdw_epsilon_to_openmm), moved verbatim.

openmm.py re-imports both sets (still used internally by _collect_*,
_build_cmap_force, OpenMMHandle, and create_context) and drops the six
now-unused q2mm.models.units imports. ~150 lines relocated; openmm.py
shrinks from 1982 to 1742 lines. The architecture doc module tree is
updated so the Phase 1 guard test stays green.
