+++ q2mm/backends/mm/openmm.py
@@ -19,6 +19,24 @@ logger = logging.getLogger(__name__)
 
 from q2mm.backends.base import MMEngine
 from q2mm.backends.registry import register_mm
+from q2mm.backends.mm._openmm_terms import (
+    _AngleTerm,
+    _BondTerm,
+    _CmapTerm,
+    _Exception14,
+    _TorsionTerm,
+    _UBTerm,
+    _VdwTerm,
+)
+from q2mm.backends.mm._openmm_units import (
+    _angle_k_to_harmonic,
+    _angle_k_to_openmm,
+    _bond_k_to_harmonic,
+    _bond_k_to_openmm,
+    _vdw_epsilon_to_openmm,
+    _vdw_radius_to_openmm,
+    _vdw_sigma_nm,
+)
 from q2mm.constants import (
     MM3_BOND_C3,
     MM3_BOND_C4,
@@ -35,16 +53,10 @@ from q2mm.models.units import (
     KCAL_TO_KJ,
     RAD_TO_DEG,
     ang_to_nm,
-    canonical_to_openmm_angle_k,
-    canonical_to_openmm_bond_k,
     canonical_to_openmm_bond_k_nm,
-    canonical_to_openmm_epsilon,
-    canonical_to_openmm_harmonic_angle_k,
-    canonical_to_openmm_harmonic_bond_k,
     canonical_to_openmm_torsion_k,
     hessian_kjmolnm2_to_au,
     kj_to_kcal,
-    rmin_half_to_sigma_nm,
 )
 from q2mm.models.forcefield import AngleParam, BondParam, ForceField, FunctionalForm, TorsionParam, VdwParam
 from q2mm.models.molecule import Q2MMMolecule
@@ -58,160 +70,6 @@ except ImportError:  # pragma: no cover - exercised when OpenMM is not installed
     mm = None
     unit = None
     _HAS_OPENMM = False
-
-
-@dataclass
-class _BondTerm:
-    """Internal record mapping a molecule bond to its OpenMM force index.
-
-    Attributes:
-        force_index: Index of this bond in the OpenMM bond force object.
-        atom_i: First atom index.
-        atom_j: Second atom index.
-        elements: Element symbols for the two atoms.
-        env_id: Chemical environment identifier for parameter matching.
-        ff_row: Row index in the source force field file, if applicable.
-
-    """
-
-    force_index: int
-    atom_i: int
-    atom_j: int
-    elements: tuple[str, str]
-    env_id: str = ""
-    ff_row: int | None = None
-
-
-@dataclass
-class _AngleTerm:
-    """Internal record mapping a molecule angle to its OpenMM force index.
-
-    Attributes:
-        force_index: Index of this angle in the OpenMM angle force object.
-        atom_i: First atom index.
-        atom_j: Central atom index.
-        atom_k: Third atom index.
-        elements: Element symbols for the three atoms.
-        env_id: Chemical environment identifier for parameter matching.
-        ff_row: Row index in the source force field file, if applicable.
-
-    """
-
-    force_index: int
-    atom_i: int
-    atom_j: int
-    atom_k: int
-    elements: tuple[str, str, str]
-    env_id: str = ""
-    ff_row: int | None = None
-
-
-@dataclass
-class _VdwTerm:
-    """Internal record mapping a molecule atom to its OpenMM vdW particle.
-
-    Attributes:
-        particle_index: Index of this particle in the OpenMM vdW force.
-        atom_type: Atom type label for parameter matching.
-        element: Element symbol.
-        ff_row: Row index in the source force field file, if applicable.
-
-    """
-
-    particle_index: int
-    atom_type: str = ""
-    element: str = ""
-    ff_row: int | None = None
-
-
-@dataclass
-class _Exception14:
-    """A 1-4 nonbonded exception whose parameters must track particle updates.
-
-    Attributes:
-        exception_index: Index of this exception in the OpenMM NonbondedForce.
-        particle_i: First particle index.
-        particle_j: Second particle index.
-
-    """
-
-    exception_index: int
-    particle_i: int
-    particle_j: int
-
-
-@dataclass
-class _TorsionTerm:
-    """Internal record mapping a molecule torsion to its OpenMM force index.
-
-    Attributes:
-        force_index: Index of this torsion in the OpenMM torsion force object.
-        atom_i: First atom index.
-        atom_j: Second atom index.
-        atom_k: Third atom index.
-        atom_l: Fourth atom index.
-        elements: Element symbols for the four atoms.
-        periodicity: Fourier component periodicity (1, 2, or 3).
-        env_id: Chemical environment identifier for parameter matching.
-        ff_row: Row index in the source force field file, if applicable.
-        is_improper: Whether this term is an improper torsion.
-
-    """
-
-    force_index: int
-    atom_i: int
-    atom_j: int
-    atom_k: int
-    atom_l: int
-    elements: tuple[str, str, str, str]
-    periodicity: int = 1
-    env_id: str = ""
-    ff_row: int | None = None
-    is_improper: bool = False
-
-
-@dataclass
-class _UBTerm:
-    """Internal record mapping a Urey-Bradley 1-3 pair to its OpenMM force index.
-
-    Attributes:
-        force_index: Index of this UB bond in the OpenMM bond force object.
-        atom_i: First atom of the angle (outer).
-        atom_k: Third atom of the angle (outer).
-        elements: Element symbols for the three angle atoms (for matching).
-        env_id: Chemical environment identifier for parameter matching.
-        ff_row: Row index in the source force field file, if applicable.
-
-    """
-
-    force_index: int
-    atom_i: int
-    atom_k: int
-    elements: tuple[str, str, str]
-    env_id: str = ""
-    ff_row: int | None = None
-
-
-@dataclass
-class _CmapTerm:
-    """Internal record mapping a CMAP correction to its OpenMM force index.
-
-    Attributes:
-        torsion_index: Index of this CMAP torsion in the CMAPTorsionForce.
-        map_index: Index of the CMAP grid in the CMAPTorsionForce.
-        phi_atoms: Atom indices for the φ dihedral (4 atoms).
-        psi_atoms: Atom indices for the ψ dihedral (4 atoms).
-        phi_types: Atom types for the φ dihedral.
-        psi_types: Atom types for the ψ dihedral.
-
-    """
-
-    torsion_index: int
-    map_index: int
-    phi_atoms: tuple[int, int, int, int]
-    psi_atoms: tuple[int, int, int, int]
-    phi_types: tuple[str, str, str, str]
-    psi_types: tuple[str, str, str, str]
 
 
 @dataclass
@@ -392,104 +250,6 @@ def _coerce_forcefield(forcefield: ForceField | None, molecule: Q2MMMolecule) ->
     if forcefield is not None:
         return forcefield
     return ForceField.create_for_molecule(molecule)
-
-
-def _bond_k_to_openmm(force_constant: float) -> float:
-    """Convert canonical bond force constant (kcal/mol/Å²) to kJ/mol/Å².
-
-    Args:
-        force_constant: Bond force constant in kcal/mol/Å².
-
-    Returns:
-        float: Bond force constant in kJ/mol/Å².
-
-    """
-    return canonical_to_openmm_bond_k(force_constant)
-
-
-def _angle_k_to_openmm(force_constant: float) -> float:
-    """Convert canonical angle force constant (kcal/mol/rad²) to kJ/mol/rad².
-
-    Args:
-        force_constant: Angle force constant in kcal/mol/rad².
-
-    Returns:
-        float: Angle force constant in kJ/mol/rad².
-
-    """
-    return canonical_to_openmm_angle_k(force_constant)
-
-
-def _bond_k_to_harmonic(force_constant: float) -> float:
-    """Canonical bond k (kcal/mol/Å²) → HarmonicBondForce k (kJ/mol/nm²).
-
-    OpenMM's HarmonicBondForce uses E = ½·k·(r−r₀)² while the canonical
-    convention is E = k·(r−r₀)², so k_openmm = 2·k.  Additionally convert
-    kcal→kJ (×4.184) and Å⁻²→nm⁻² (×100).
-
-    Args:
-        force_constant: Bond force constant in kcal/mol/Å².
-
-    Returns:
-        float: Bond force constant in kJ/mol/nm² with the ½ convention.
-
-    """
-    return canonical_to_openmm_harmonic_bond_k(force_constant)
-
-
-def _angle_k_to_harmonic(force_constant: float) -> float:
-    """Canonical angle k (kcal/mol/rad²) → HarmonicAngleForce k (kJ/mol/rad²).
-
-    OpenMM's HarmonicAngleForce uses E = ½·k·(θ−θ₀)² while the canonical
-    convention is E = k·(θ−θ₀)², so k_openmm = 2·k.  Convert kcal→kJ.
-
-    Args:
-        force_constant: Angle force constant in kcal/mol/rad².
-
-    Returns:
-        float: Angle force constant in kJ/mol/rad² with the ½ convention.
-
-    """
-    return canonical_to_openmm_harmonic_angle_k(force_constant)
-
-
-def _vdw_sigma_nm(radius: float) -> float:
-    """Convert Rmin/2 (Å) to LJ sigma (nm) for standard 12-6 NonbondedForce.
-
-    Args:
-        radius: Van der Waals radius (Rmin/2) in Å.
-
-    Returns:
-        float: LJ sigma in nm.
-
-    """
-    return rmin_half_to_sigma_nm(radius)
-
-
-def _vdw_radius_to_openmm(radius: float) -> float:
-    """Convert vdW radius from Å to nm for CustomNonbondedForce.
-
-    Args:
-        radius: Van der Waals radius in Å.
-
-    Returns:
-        float: Van der Waals radius in nm.
-
-    """
-    return ang_to_nm(radius)
-
-
-def _vdw_epsilon_to_openmm(epsilon: float) -> float:
-    """Convert vdW epsilon from kcal/mol to kJ/mol.
-
-    Args:
-        epsilon: Well depth in kcal/mol.
-
-    Returns:
-        float: Well depth in kJ/mol.
-
-    """
-    return canonical_to_openmm_epsilon(epsilon)
 
 
 def _collect_bond_assignments(
